requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# OCR
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1

# Ollama
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Tesseract OCR
pytesseract==0.3.10

# NOTE: Folgende Dependencies sind entfernt (vorerst nicht verwendet):
# - rdflib, sparqlwrapper (Triple Store vorerst nicht verwendet)
# - wikibaseintegrator (Wikidata vorerst nicht verwendet)
# - spacy (NLP entfernt)
